Rename and replace
