+++ src/inscription.rs
@@ -1,6 +1,5 @@
 pub mod brc20;
 pub mod nft;
-pub mod parser;
 
 use bitcoin::script::{Builder as ScriptBuilder, PushBytesBuf};
 use serde::de::DeserializeOwned;
